forward_auth_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyjwt==2.8.0
redis==5.0.1
httpx==0.25.1
python-multipart==0.0.6
python-dotenv==1.0.0
prometheus-client==0.18.0
cryptography==41.0.7